requirements.txt
pytorch
torchvision


# clip
ftfy 
regex 
tqdm
openai-clip
